Refactor CommitObserver to a separate file
